refactor(pk): extract pk/conventionalcommits sub-package
